# integrations/al-gate/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "AL Mnemo Gate", "version": "1.0.0"},
  "servers": [{"url": "REPLACE_WITH_TAILSCALE_FUNNEL_URL"}],
  "paths": {
    "/recall": {
      "post": {
        "operationId": "recallMemory",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Recall"}}}},
        "responses": {"200": {"description": "Matching AL memories"}}
      }
    },
    "/save": {
      "post": {
        "operationId": "saveMemory",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Save"}}}},
        "responses": {"200": {"description": "Memory saved"}}
      }
    }
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}},
    "schemas": {
      "Recall": {
        "type": "object", "additionalProperties": false, "required": ["prompt"],
        "properties": {
          "prompt": {"type": "string", "maxLength": 4000},
          "max_results": {"type": "integer", "minimum": 1, "maximum": 10, "default": 5},
          "category": {"type": "string", "enum": ["session_log", "idea", "decision", "identity", "relationship"]},
          "mode": {"type": "string", "enum": ["focus", "explore"], "default": "focus"}
        }
      },
      "Save": {
        "type": "object", "additionalProperties": false,
        "required": ["session_id", "summary", "category"],
        "properties": {
          "session_id": {"type": "string", "maxLength": 128},
          "summary": {"type": "string", "maxLength": 6000},
          "key_facts": {"type": "array", "maxItems": 30, "items": {"type": "string"}},
          "category": {"type": "string", "enum": ["session_log", "idea", "decision", "identity", "relationship"]},
          "additional_tags": {"type": "array", "maxItems": 20, "items": {"type": "string"}}
        }
      }
    }
  }
}